extend serialization unitttest
